Avoid deprecated wxPen/wxBrush/wxFont API in wxX11 code.
Also simplify the code by relying on implicit constructors of wxPen and
wxBrush from wxColour.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@77873 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/x11/font.cpp
@@ -182,8 +182,8 @@ void wxFontRefData::Init(int pointSize,
     m_faceName = faceName;
 
     // we accept both wxDEFAULT and wxNORMAL here - should we?
-    m_style = style == wxDEFAULT ? wxFONTSTYLE_NORMAL : style;
-    m_weight = weight == wxDEFAULT ? wxFONTWEIGHT_NORMAL : weight;
+    m_style = static_cast<int>(style) == wxDEFAULT ? wxFONTSTYLE_NORMAL : style;
+    m_weight = static_cast<int>(weight) == wxDEFAULT ? wxFONTWEIGHT_NORMAL : weight;
 
     m_underlined = underlined;
     m_strikethrough = strikethrough;
